+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/constant/AnimationConstant.kt
@@ -1,17 +1,14 @@
 package net.mullvad.mullvadvpn.constant
 
-import androidx.compose.animation.core.Spring
 import net.mullvad.mullvadvpn.lib.model.LatLong
 import net.mullvad.mullvadvpn.lib.model.Latitude
 import net.mullvad.mullvadvpn.lib.model.Longitude
 
 const val MINIMUM_LOADING_TIME_MILLIS = 500L
 
-const val SCREEN_ANIMATION_TIME_MILLIS = Spring.StiffnessMediumLow.toInt()
-
-const val HORIZONTAL_SLIDE_FACTOR = 1 / 3f
-
-fun Int.withHorizontalScalingFactor(): Int = (this * HORIZONTAL_SLIDE_FACTOR).toInt()
+const val ENTER_TRANSITION_SCALE_IN_FACTOR = 1.1f
+const val ENTER_TRANSITION_SLIDE_FACTOR = 0.25f
+const val EXIT_TRANSITION_SLIDE_FACTOR = 0.1f
 
 const val SECURE_ZOOM = 1.15f
 const val UNSECURE_ZOOM = 1.20f